+++ crates/web-app-dioxus/src/page/home.rs
@@ -90,30 +90,18 @@ pub fn Home() -> Element {
                 }
             })
         }
-        CacheState::Error(
-            domain::ReadError::NotFound
-            | domain::ReadError::Storage(domain::StorageError::NoConnection),
-        ) => None,
         CacheState::Error(err) => Some(rsx! { Error { "{err}" } }),
         CacheState::Loading => Some(rsx! { Loading {} }),
     };
 
     let routines_subtitle = match &*cache.routines.read() {
-        CacheState::Ready(_)
-        | CacheState::Error(
-            domain::ReadError::NotFound
-            | domain::ReadError::Storage(domain::StorageError::NoConnection),
-        ) => None,
+        CacheState::Ready(_) => None,
         CacheState::Error(err) => Some(rsx! { Error { "{err}" } }),
         CacheState::Loading => Some(rsx! { Loading {} }),
     };
 
     let exercises_subtitle = match &*cache.exercises.read() {
-        CacheState::Ready(_)
-        | CacheState::Error(
-            domain::ReadError::NotFound
-            | domain::ReadError::Storage(domain::StorageError::NoConnection),
-        ) => None,
+        CacheState::Ready(_) => None,
         CacheState::Error(err) => Some(rsx! { Error { "{err}" } }),
         CacheState::Loading => Some(rsx! { Loading {} }),
     };
@@ -124,10 +112,6 @@ pub fn Home() -> Element {
             .filter(|bw| bw.date <= today)
             .max_by(|a, b| a.date.cmp(&b.date))
             .map(|bw| rsx! { strong { "{bw.weight:.1} kg" } " ({last(bw.date)})" }),
-        CacheState::Error(
-            domain::ReadError::NotFound
-            | domain::ReadError::Storage(domain::StorageError::NoConnection),
-        ) => None,
         CacheState::Error(err) => Some(rsx! { Error { "{err}" } }),
         CacheState::Loading => Some(rsx! { Loading {} }),
     };
@@ -141,10 +125,6 @@ pub fn Home() -> Element {
                 bf.jp3(sex())
                     .map(|jp3| rsx! { strong { "{jp3:.1} %" } " ({last(bf.date)})" })
             }),
-        CacheState::Error(
-            domain::ReadError::NotFound
-            | domain::ReadError::Storage(domain::StorageError::NoConnection),
-        ) => None,
         CacheState::Error(err) => Some(rsx! { Error { "{err}" } }),
         CacheState::Loading => Some(rsx! { Loading {} }),
     };
@@ -153,20 +133,6 @@ pub fn Home() -> Element {
         match (&*cache.body_fat.read(), &*cache.body_weight.read()) {
             (CacheState::Ready(_), CacheState::Ready(_)) => latest_ffmi()
                 .map(|(date, value)| rsx! { strong { "{value:.1}" } " ({last(date)})" }),
-            (
-                CacheState::Error(
-                    domain::ReadError::NotFound
-                    | domain::ReadError::Storage(domain::StorageError::NoConnection),
-                ),
-                _,
-            )
-            | (
-                _,
-                CacheState::Error(
-                    domain::ReadError::NotFound
-                    | domain::ReadError::Storage(domain::StorageError::NoConnection),
-                ),
-            ) => None,
             (CacheState::Error(err), _) | (_, CacheState::Error(err)) => {
                 Some(rsx! { Error { "{err}" } })
             }
@@ -182,10 +148,6 @@ pub fn Home() -> Element {
                 CacheState::Ready(period) => domain::current_cycle(&domain::cycles(period)).map(|current_cycle| rsx! {
                     strong { "{current_cycle.time_left.num_days()} (±{current_cycle.time_left_variation.num_days()}) days left" } " (day {(today - current_cycle.begin).num_days()})"
                 }),
-                CacheState::Error(
-                    domain::ReadError::NotFound
-                    | domain::ReadError::Storage(domain::StorageError::NoConnection),
-                ) => None,
                 CacheState::Error(err) => Some(rsx! { Error { "{err}" } }),
                 CacheState::Loading => Some(rsx! { Loading {} }),
             }